Module nnet::params

Implemented parameters for neural nets or trainers.

Structs

CEFunction
Cross Entropy error function.

DefaultErrorGradient
Default Error Gradient functions.

DefaultWeightFunction
Default weight function that is dependent on the input size.

LogisticNeuralNet
Default Parameters for a Logistic Neural Net.

MSEFunction
MSE Error function.

NegativeOneBiasFunction
Returns -1 for each bias node.

PositiveOneBiasFunction
Returns 1 for each bias node.

RandomBiasWeightFunction
Bias function that returns a random weight between -0.5 and 0.5.

TanhNeuralNet
Default Parameters for a Tanh Neural Net.