Trait nnet::prelude::NeuralNet

pub trait NeuralNet<P> where P: NeuralNetParameters {
    fn dim_input() -> usize;
    fn dim_output() -> usize;
    fn dim_hidden() -> usize;
    fn node(&self, i: Node) -> f64;
    fn node_mut(&mut self, i: Node) -> &mut f64;
    fn layer(&self, layer: Layer) -> &[f64];
    fn predict(&mut self, inp: &[f64]);
}

A single-layer neural network.

Required Methods

fn dim_input() -> usize

Returns the dimensions of the input layer.

fn dim_output() -> usize

Returns the dimensions of the output layer.

fn dim_hidden() -> usize

Returns the dimensions of the hidden layer.

fn node(&self, i: Node) -> f64

Returns the value of a node in a layer at a specified coordinate.

fn node_mut(&mut self, i: Node) -> &mut f64

Returns a mutable reference to a node in a layer.

fn layer(&self, layer: Layer) -> &[f64]

Returns the specified layer.

fn predict(&mut self, inp: &[f64])

Computes the predicted value for a given input and stores it internally. The prediction can be retrieved using layer. The reason predict doesn't return the prediction, is because it requires a mutable borrow on self.