Module nnet::trainer::backpropagation

Implementation of backpropagation trainers.

Modules

parallel
Multithreaded implementations of backpropagation trainers.

Structs

BatchEpochTrainer
Back-propagation trainer where the stopping condition is based on a max number of epochs. Weights are updated at the end of each epoch.

SeqEpochTrainer
Back-propagation trainer where the stopping criteria is bounded by the epoch. Weights are updated for each example in the training set.

SeqErrorAverageTrainer
Back-propagation trainer where the stopping condition is primarily the calculated average error, with an optional stopping condition based on the epoch. Weights are updated for each example in the training set.